fixed some tests

+++ pkg/user/profile_store.go
@@ -108,7 +108,7 @@ func (s *MySQLStore) BulkCreateProfile(ctx context.Context, profiles []Profile) 
 	return profiles, nil
 }
 
-func (s *MySQLStore) FetchProfileByUserID(ctx context.Context, id uint32) (profile Profile, err error) {
+func (s *MySQLStore) FetchProfileByUserID(ctx context.Context, id int64) (profile Profile, err error) {
 	return s.fetchProfileByQuery(ctx, "SELECT * FROM `user_profile` WHERE user_id = ? LIMIT 1", id)
 }
 
@@ -155,7 +155,7 @@ func (s *MySQLStore) UpdateProfile(ctx context.Context, profile Profile, changel
 	return profile, nil
 }
 
-func (s *MySQLStore) DeleteProfileByUserID(ctx context.Context, userID uint32) (err error) {
+func (s *MySQLStore) DeleteProfileByUserID(ctx context.Context, userID int64) (err error) {
 	if userID == 0 {
 		return ErrZeroID
 	}
